While preparing to redo loading of logconfig, accidentally rewrite resource manager's search path handling.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@894 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/ShipEntity.m
@@ -8284,7 +8284,7 @@ static void ApplyConstantUniforms(NSDictionary *uniforms, GLhandleARB shaderProg
 				glUniform1fARB(variableLocation, floatValue);
 			}
 		}
-		else if ([type isEqualToString:@"float"])
+		else if ([type isEqualToString:@"int"])
 		{
 			if ([value respondsToSelector:@selector(intValue)])
 			{
